refactor api
refactoring, moving api directory from /imports/ui to /imports. Additionally creating /imports/api/users directory for methods.ts
