+++ finding-nibbles-ts/client/main.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Meteor } from 'meteor/meteor';
 import { App } from '/imports/ui/App';
-import '/imports/ui/api/searchHistory';
-import '../imports/ui/api/SavedRestaurants';
-import { SavedRestaurantsCollection } from '../imports/ui/api/SavedRestaurants';
+import '../imports/api/searchHistory';
+import '../imports/api/SavedRestaurants';
+import { SavedRestaurantsCollection } from '../imports/api/SavedRestaurants';
 
 // Expose to window for debugging
 (window as any).SavedRestaurantsCollection = SavedRestaurantsCollection;
