+++ finding-nibbles-ts/imports/api/generateImage.ts
