+++ finding-nibbles-ts/imports/api/initialFoodData.ts
@@ -1,4 +1,4 @@
-import {FoodItems} from '../../../server/FoodItems';
+import {FoodItems} from '../../server/FoodItems';
 
 const initialFoods = [
     {
